Remove Unneeded Items and Disable Ask a Friend with No Subjects
