+++ GalaxyZooTouchTable/GalaxyZooTouchTable/GalaxyZooTouchTable.csproj
@@ -183,7 +183,6 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Behaviors\UIElementDragBehavior.cs" />
-    <Compile Include="Behaviors\UIElementDropAdorner.cs" />
     <Compile Include="Behaviors\UIElementDropBehavior.cs" />
     <Compile Include="Converters\EqualityConverter.cs" />
     <Compile Include="Converters\FontSizeConverter.cs" />
@@ -195,6 +194,7 @@
     <Compile Include="Lib\ClassifierViewEnum.cs" />
     <Compile Include="Lib\GlobalData.cs" />
     <Compile Include="Lib\ContainerHelper.cs" />
+    <Compile Include="Lib\GalaxyAdorner.cs" />
     <Compile Include="Lib\GridLengthAnimation.cs" />
     <Compile Include="Lib\HandlingEventTrigger.cs" />
     <Compile Include="Lib\NotificationStatus.cs" />
