+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Views/DropZone.xaml
@@ -22,31 +22,33 @@
         <i:Interaction.Behaviors>
             <Behaviors:UIElementDropBehavior/>
         </i:Interaction.Behaviors>
-        <Border Width="106" Height="106" BorderThickness="2" CornerRadius="3">
-            <Border.BorderBrush>
-                <DrawingBrush Viewport="0,0,8,8" ViewportUnits="Absolute" TileMode="Tile">
-                    <DrawingBrush.Drawing>
-                        <DrawingGroup>
-                            <GeometryDrawing Brush="{StaticResource AttentionColor}">
-                                <GeometryDrawing.Geometry>
-                                    <GeometryGroup>
-                                        <RectangleGeometry Rect="0,0,50,50"/>
-                                        <RectangleGeometry Rect="50,50,50,50"/>
-                                    </GeometryGroup>
-                                </GeometryDrawing.Geometry>
-                            </GeometryDrawing>
-                        </DrawingGroup>
-                    </DrawingBrush.Drawing>
-                </DrawingBrush>
-            </Border.BorderBrush>
-            <TextBlock
-                Text="Drag and drop a galaxy here to get started."
-                TextWrapping="Wrap"
-                Width="80"
-                Height="58"
-                Foreground="White"
-                FontFamily="/GalaxyZooTouchTable;component/Fonts/#Karla"
-                FontWeight="Bold"/>
+            <Border>
+                <Border Width="106" Height="106" BorderThickness="2" CornerRadius="3">
+                <Border.BorderBrush>
+                    <DrawingBrush Viewport="0,0,8,8" ViewportUnits="Absolute" TileMode="Tile">
+                        <DrawingBrush.Drawing>
+                            <DrawingGroup>
+                                <GeometryDrawing Brush="{StaticResource AttentionColor}">
+                                    <GeometryDrawing.Geometry>
+                                        <GeometryGroup>
+                                            <RectangleGeometry Rect="0,0,50,50"/>
+                                            <RectangleGeometry Rect="50,50,50,50"/>
+                                        </GeometryGroup>
+                                    </GeometryDrawing.Geometry>
+                                </GeometryDrawing>
+                            </DrawingGroup>
+                        </DrawingBrush.Drawing>
+                    </DrawingBrush>
+                </Border.BorderBrush>
+                <TextBlock
+                    Text="Drag and drop a galaxy here to get started."
+                    TextWrapping="Wrap"
+                    Width="80"
+                    Height="58"
+                    Foreground="White"
+                    FontFamily="/GalaxyZooTouchTable;component/Fonts/#Karla"
+                    FontWeight="Bold"/>
+            </Border>
         </Border>
     </Border>
 </UserControl>
